chore: Add custom token substrate pallet code
