--- guessing-random-number-substrate-custom-blockchain/pallets/template/README.md
@@ -1,1 +1,0 @@
-License: MIT-0
