--- guessing-random-number-substrate-custom-blockchain/runtime/build.rs
@@ -1,10 +1,0 @@
-fn main() {
-	#[cfg(feature = "std")]
-	{
-		substrate_wasm_builder::WasmBuilder::new()
-			.with_current_project()
-			.export_heap_base()
-			.import_memory()
-			.build();
-	}
-}
