--- guessing-random-number-substrate-custom-blockchain/rust-toolchain.toml
@@ -1,14 +1,0 @@
-[toolchain]
-channel = "nightly"
-components = [
-	"cargo",
-	"clippy",
-	"rust-analyzer",
-	"rust-src",
-	"rust-std",
-	"rustc-dev",
-	"rustc",
-	"rustfmt",
-]
-targets = [ "wasm32-unknown-unknown" ]
-profile = "minimal"
