+++ substrate-blockchain/guessing-random-number-game/node/src/main.rs
@@ -10,5 +10,5 @@ mod command;
 mod rpc;
 
 fn main() -> sc_cli::Result<()> {
-	command::run()
+    command::run()
 }
