+++ substrate-blockchain/guessing-random-number-game/runtime/build.rs
@@ -1,0 +1,10 @@
+fn main() {
+    #[cfg(feature = "std")]
+    {
+        substrate_wasm_builder::WasmBuilder::new()
+            .with_current_project()
+            .export_heap_base()
+            .import_memory()
+            .build();
+    }
+}
